iop_rust_tests::bindings

Function iop_obj_get_field_type

pub unsafe extern "C" fn iop_obj_get_field_type(
    iop_env_ctx: *const iop_env_ctx_t,
    st: *const iop_struct_t,
    value: *const c_void,
    path: lstr_t,
    type_: *mut iop_full_type_t,
    is_array: *mut bool,
    err: *mut sb_t,
) -> i32

Get the type of a field for a given IOP object.

This function differs from \ref t_iop_field_path_compile + \ref iop_field_path_get_type in that it applies to a specific IOP object, and is not generic on any object of a given type. This allows the path to use fields from subclasses used by this IOP object. For example, the path “arrayOfParentClass[1].fieldOfSubclass” cannot be used with t_iop_field_path_compile. However, if passed to this function, it will be able to resolve the type, if the given IOP object has the right subclass in this path.

\param[in] iop_env_ctx The current IOP environment context. \param[in] st Type of the IOP object. \param[in] value Pointer to the IOP object. \param[in] path Path to the IOP field. See \ref mp_iop_field_path_compile for the syntax. \param[out] type Type of the IOP field for the given IOP object. \param[out] is_array True if the field is an array. \param[out] err The error description in case of error.

\return -1 In case of error, 0 otherwise.