iop_rust_tests::bindings

Function mp_iop_dup_desc_flags_sz

pub unsafe extern "C" fn mp_iop_dup_desc_flags_sz(
    mp: *mut mem_pool_t,
    st: *const iop_struct_t,
    v: *const c_void,
    flags: u32,
    sz: *mut usize,
) -> *mut c_void

Duplicate an IOP structure.

The resulting IOP structure will fully contained in one block of memory.

Prefer the macro versions instead of this low-level API.

\param[in] mp The memory pool to use for the new allocation. If mp is NULL the libc malloc() will be used. \param[in] st The IOP structure definition (__s). \param[in] v The IOP structure to duplicate. \param[in] flags The bitmap of \ref iop_copy_flags_t. \param[out] sz If set, filled with the size of the allocated buffer.