iop_rust_tests::bindings

Function mpsc_queue_drain_fast

pub unsafe extern "C" fn mpsc_queue_drain_fast(
    it: *mut mpsc_it_t,
    doit: Option<unsafe extern "C" fn(*mut mpsc_node_t, data_t)>,
    data: data_t,
) -> *mut mpsc_node_t

\brief fast path of the drain.

\param[in] it the iterator to use \param[in] doit a function or macro that takes one mpsc_node_t as an argument: the node to process. \param[in] data additionnal data to pass to \p doit.

\returns a non NULL mpsc_node_t that may be the last one in the queue for processing purposes. mpsc_queue_drain_fast doesn’t process this node, its up to the caller to do it. This node should NOT be freed, as mpsc_queue_drain_end will need it.