iop_rust_tests::bindings

Function thr_syn_wait_flags

pub unsafe extern "C" fn thr_syn_wait_flags(
    syn: *mut thr_syn_t,
    flags: u32,
)

Wait for the completion of a given macro task.

This function waits until all jobs that have been registered as synchronizing against this \ref thr_syn_t to complete. If there are jobs running, the waiter is used to perform new jobs in between to avoid blocking.

\param[in] syn A \ref thr_syn_t on which you wish to wait. \param[in] flags An optional combination of \ref thr_syn_wait_flags_t flags to alter the behavior of the waiter.