iop_rust_tests::bindings

Function write_in_tmp_file

pub unsafe extern "C" fn write_in_tmp_file(
    file_path: *mut i8,
    data: *const i8,
    len: i32,
    err: *mut sb_t,
) -> i32

Create a temporary file and write data in it.

Standard ‘unlink’ method should be called on the path given in argument in case of success (once you no longer need the file).

\param[in] file_path path of the file. It will create with the method mkstemp. As mentioned, in the doc of this method, the last six characters of template must be “XXXXXX” and these are replaced with a string that makes the filename unique. \param[in] data the data to write. \param[in] len length of data. \param[out] err error on creating file or writing data.

\return 0 on success, -1 the data has not be written.